fix(ui): add role description docs
