+++ ui/src/locales/en/common.json
@@ -44,5 +44,6 @@
     "nameStartChar": "Must start with a letter or digit",
     "nameInvalidChars": "Can only contain lowercase letters, digits, dots, underscores, and hyphens",
     "minNumber": "{{field}} must be greater than or equal to {{min}}"
-  }
+  },
+  "docs": "https://matrixhub.ai{{doc}}"
 }
