+++ ui/src/locales/zh/common.json
@@ -44,5 +44,6 @@
     "nameStartChar": "必须以字母或数字开头",
     "nameInvalidChars": "只能包含小写字母、数字、点、下划线和连字符",
     "minNumber": "{{field}}必须大于或等于{{min}}"
-  }
+  },
+  "docs": "https://matrixhub.ai/zh-cn{{doc}}"
 }
